Bug 1209639 - Fix warning C4146: unary minus operator applied to unsigned type, result still unsigned. r=rkent a=aleth,neil SM CLOSED TREE
--HG--
extra : amend_source : e7b83a8c2a105df9022ede232139d1a9583d0e58

+++ mailnews/news/src/nsNntpIncomingServer.cpp
@@ -1657,7 +1657,7 @@ nsNntpIncomingServer::SetSearchValue(const nsAString &aSearchValue)
 
   if (mTree) {
     mTree->BeginUpdateBatch();
-    mTree->RowCountChanged(0, -mSubscribeSearchResult.Length());
+    mTree->RowCountChanged(0, -static_cast<int32_t>(mSubscribeSearchResult.Length()));
   }
 
   nsTArray<nsCString> searchStringParts;
